refactor: src/utils/edtr-io.ts -> src/utils/serlo-editor.ts
