+++ src/utils/create-migration.ts
@@ -1,6 +1,6 @@
 import { ApiCache } from './api-cache'
 import { CallbackBasedDatabase, createDatabase, Database } from './database'
-import { isPlugin } from './edtr-io'
+import { isPlugin } from './serlo-editor'
 import { SlackLogger } from './slack-logger'
 
 export function createMigration(
