+++ src/utils/index.ts
@@ -2,6 +2,6 @@ export { ApiCache } from './api-cache'
 export * from './create-migration'
 export * from './database'
 export * from './delete-utils'
-export * from './edtr-io'
+export * from './serlo-editor'
 export * from './sql-utils'
 export * from './slack-logger'
